Changed location of OwnerResolver and StadiumResolver for architectural purposes
